specificItem
