+++ dev/Item.java
@@ -4,44 +4,24 @@ import java.time.Instant;
 public class Item {
     private String name;
     private String catalogNum;
-    private final Date expirationDate;
     private double weight;
     private String catalogName;
     private int minQuantity;
     private int amount = 0;
     private TempLevel temperature;
-    private Location location;
-    private boolean isDefected = false;
-    private int specificId;
 
 
-
-
-    public Item(String name, String catalogNum, Date expirationDate, double weight, String catalogName, TempLevel temperature, int minQuantity, Location location)
+    public Item(String name, String catalogNum, double weight, String catalogName, TempLevel temperature, int minQuantity)
     {
         this.name = name;
         this.catalogNum = catalogNum;
-        this.expirationDate = expirationDate;
         this.weight = weight;
         this.catalogName = catalogName;
         this.temperature = temperature;
         this.minQuantity = minQuantity;
-        this.location = location;
 
     }
 
-
-    public void setDefected(boolean defected) {
-        isDefected = defected;
-    }
-
-    public Location getLocation() {
-        return location;
-    }
-
-    public void setLocation(Location location) {
-        this.location = location;
-    }
 
     public void setAmount(int amount) {
         this.amount = amount;
@@ -61,20 +41,6 @@ public class Item {
 
     public int getAmount() {
         return amount;
-    }
-
-    public boolean isExpired()
-    {
-        return expirationDate.before(Date.from(Instant.now()));
-    }
-
-    public boolean isDefected()
-    {
-        if (isExpired() == true)
-        {
-            isDefected = true;
-        }
-        return isDefected;
     }
